+++ documentation/changelog.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## Viikko 3
+## Week 3
 - Created the Minesweeper Class which will be the basis for the game logic
 - Added the functionality of creating the mines into the game
 - Added the functionality of revealing a tile
 - Tested the mine creation functionality
+
+## Week 4
+- Added functionality of initializing the game board
+- Each cell now has properties to make them easier to manage
+- Functionalities of revealing and flagging cells have been developed
+- Added a text based display as a placeholder